requirements.txt
# Core dependencies
numpy>=1.21.0
tensorflow>=2.10.0
scikit-learn>=1.0.0
joblib>=1.0.0

# Audio processing
sounddevice>=0.4.0
scipy>=1.7.0

# Serial communication
pyserial>=3.5

# Optional: for web dashboard
# websockets>=10.0
# fastapi>=0.95.0
# uvicorn>=0.20.0
